feat(storage): introduced browser-storage as common super-class
localStorage and sessionStorage both implement the window.Storage interface

By passing the desired storage into browser-storage we avoid duplicating code
