+++ packages/analytics-browser/test/storage/local-storage.test.ts
@@ -2,79 +2,14 @@ import { LocalStorage } from '../../src/storage/local-storage';
 import * as AnalyticsClientCommon from '@amplitude/analytics-client-common';
 
 describe('local-storage', () => {
-  describe('isEnabled', () => {
-    test('should return true', async () => {
-      const localStorage = new LocalStorage();
-      expect(await localStorage.isEnabled()).toBe(true);
-    });
+  test('should return true if storage is available', async () => {
+    const localStorage = new LocalStorage();
+    expect(await localStorage.isEnabled()).toBe(true);
   });
 
-  describe('get', () => {
-    test('should return undefined if not set', async () => {
-      const localStorage = new LocalStorage();
-      expect(await localStorage.get('1')).toBe(undefined);
-    });
-
-    test('should return object', async () => {
-      const localStorage = new LocalStorage<Record<string, number>>();
-      await localStorage.set('1', { a: 1 });
-      expect(await localStorage.get('1')).toEqual({ a: 1 });
-    });
-
-    test('should return array', async () => {
-      const localStorage = new LocalStorage<number[]>();
-      await localStorage.set('1', [1]);
-      expect(await localStorage.get('1')).toEqual([1]);
-    });
-  });
-
-  describe('set', () => {
-    test('should set value', async () => {
-      const localStorage = new LocalStorage();
-      await localStorage.set('1', 'a');
-      expect(await localStorage.get('1')).toBe('a');
-    });
-  });
-
-  describe('remove', () => {
-    test('should remove value of key', async () => {
-      const localStorage = new LocalStorage();
-      await localStorage.set('1', 'a');
-      await localStorage.set('2', 'b');
-      expect(await localStorage.get('1')).toBe('a');
-      expect(await localStorage.get('2')).toBe('b');
-      await localStorage.remove('1');
-      expect(await localStorage.get('1')).toBe(undefined);
-      expect(await localStorage.get('2')).toBe('b');
-    });
-
-    test('should handle when GlobalScope is not defined', async () => {
-      const localStorage = new LocalStorage<number[]>();
-      jest.spyOn(AnalyticsClientCommon, 'getGlobalScope').mockReturnValue(undefined);
-      await localStorage.set('1', [1]);
-      expect(await localStorage.get('1')).toEqual(undefined);
-      await localStorage.remove('1');
-    });
-  });
-
-  describe('reset', () => {
-    test('should remove all values', async () => {
-      const localStorage = new LocalStorage();
-      await localStorage.set('1', 'a');
-      await localStorage.set('2', 'b');
-      expect(await localStorage.get('1')).toBe('a');
-      expect(await localStorage.get('2')).toBe('b');
-      await localStorage.reset();
-      expect(await localStorage.get('1')).toBe(undefined);
-      expect(await localStorage.get('2')).toBe(undefined);
-    });
-
-    test('should handle when GlobalScope is not defined', async () => {
-      const localStorage = new LocalStorage<number[]>();
-      jest.spyOn(AnalyticsClientCommon, 'getGlobalScope').mockReturnValue(undefined);
-      await localStorage.set('1', [1]);
-      expect(await localStorage.get('1')).toEqual(undefined);
-      await localStorage.reset();
-    });
+  test('should return false if storage is unavailable', async () => {
+    jest.spyOn(AnalyticsClientCommon, 'getGlobalScope').mockReturnValue(undefined);
+    const localStorage = new LocalStorage();
+    expect(await localStorage.isEnabled()).toBe(false);
   });
 });
